// TodoApp-TS/src/type.d.ts
import { TODO_FILTERS } from "./const";

export interface Todo {
  id: string;
  title: string;
  completed: boolean;
}

export type Todoid = Pick<Todo, "id">;
export type TodoTitle = Pick<Todo, "title">;
export type TodoCompleted = Pick<Todo, "completed">;

export type ListOfTodos = Todo[];

export type FilterValue = (typeof TODO_FILTERS)[keyof typeof TODO_FILTERS];